[planning] continuous_collision: fit tests to CI budgets across build flavors
The suite was written against Release timings and a Release corpus, so
several targets either failed or overran under the instrumented build
flavors. Each fix is local to the flavor that needed it.

Timing claims out of concurrency_test. DeepWorkloadIsFasterInParallel and
SmallCheckIsNotSlowerInParallel are wall-clock claims in a file that is
otherwise all equalities. Valgrind serializes threads, which inverts
"parallel is faster than serial" and fails them for a reason unrelated to
the driver. They move to concurrency_timing_test, which carries
disable_in_compilation_mode_dbg and no_valgrind_tools; the determinism and
invariance cases stay in concurrency_test and keep running everywhere,
sanitizers included. The corpus, the deep workload and the option
defaults they share now live in test/concurrency_test_utilities.h. The
skip predicate also consults VALGRIND_OPTS, ASAN_OPTIONS, LSAN_OPTIONS,
TSAN_OPTIONS and UBSAN_OPTIONS, since a tool that instruments at runtime
is invisible to the compile-time tests it used to rely on (the same check
limit_malloc.cc and gcs_trajectory_optimization_test.cc make).

soundness_fuzz_test. Corpus composition was asserted as absolute counts,
which pinned the test to a 200-case corpus. The floors become fractions
of kNumCases, and kNumCases drops to 50 under sanitizers or memcheck,
where the dense cross-check is one to two orders of magnitude slower. The
floors still demand at least one case of every kind at that size, which a
static_assert enforces. asan and lsan are excluded outright (the corpus
is deliberately leaked), the timeout goes to "long", and the file's
budget note now says which margins are Release-only.

Smaller items: motion_bound_test gains a moderate timeout; certifier_test
declares num_threads = 8 for the eight caller threads it spawns; the two
temp_directory() calls and all three unchecked std::ofstream writes are
gone, since the cube mesh is now Drake's shipped geometry/test/
quad_cube.obj scaled to the same half-extents through the non-uniform
Mesh/Convex scale argument, and both L-prisms -- which encode analytic
expectations no shipped asset matches -- are built as InMemoryMesh.

The benchmark loses its manual tag and becomes testonly so CI compiles
it, defaults --out to TEST_TMPDIR, and gains a two-second smoke test on
the cheapest scenario.

Verified: full suite green; concurrency_test clean under --config=tsan
(concurrency_timing_test is filtered out there by its own no_tsan tag, as
intended); soundness_fuzz_test passes its fractional assertions with the
shrunk corpus forced on under both defines.
